requirements.txt
pyyaml
pyarmor